tools/clean-svg: don't force the group ID onto sub-elements
The theory (I think) here is that we can have a group "A" that
then applies to each subsequent element automatically without having
to indidivually name those elements.

This is both buggy and unnecessary and it completely breaks SVG files
generated by inkscape (1.4 at least) which will automatically assign a
group ID like "g1" to the group - resulting in our buttons etc. all
labeled as "Buttong1".

Remove this "feature".